module: 각 모듈 내부의 di 모듈을 상위 모듈과 통합
본디 di 모듈을 따로 두었던 것은 상위 모듈을 pure kotlin 모듈로 만들기 위함이었지만, 모든 모듈이 android 모듈인 상황에서 di 모듈을 따로 분리할 필요가 없다.
