--- database/di/src/main/java/com/practice/di/meal/MealDataSourceModule.kt
@@ -1,20 +1,0 @@
-package com.practice.di.meal
-
-import com.practice.database.meal.MealDataSource
-import com.practice.database.meal.LocalMealDataSource
-import com.practice.database.meal.room.MealDao
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-
-@Module
-@InstallIn(SingletonComponent::class)
-object MealDataSourceModule {
-
-    @Provides
-    fun provideLocalMealDataSource(
-        mealDao: MealDao
-    ): MealDataSource = LocalMealDataSource(mealDao)
-
-}
