+++ domain/build.gradle.kts
@@ -1,6 +1,8 @@
 plugins {
     id("com.android.library")
     id("org.jetbrains.kotlin.android")
+    id("kotlin-kapt")
+    id("dagger.hilt.android.plugin")
 }
 
 android {
@@ -11,10 +13,42 @@ android {
         minSdk = 23
         targetSdk = 33
 
-        testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner = "com.example.domain.di.HiltTestRunner"
         consumerProguardFiles.add(File("consumer-rules.pro"))
         testOptions {
-            unitTests.isReturnDefaultValues = true
+            unitTests {
+                isReturnDefaultValues = true
+            }
+            managedDevices {
+                devices {
+                    maybeCreate<com.android.build.api.dsl.ManagedVirtualDevice>("pixel2api30atd").apply {
+                        device = "Pixel 2"
+                        apiLevel = 30
+                        systemImageSource = "aosp-atd"
+                    }
+                }
+                devices {
+                    maybeCreate<com.android.build.api.dsl.ManagedVirtualDevice>("pixel2api30").apply {
+                        device = "Pixel 2"
+                        apiLevel = 30
+                        systemImageSource = "aosp"
+                    }
+                }
+                devices {
+                    maybeCreate<com.android.build.api.dsl.ManagedVirtualDevice>("pixel2api27").apply {
+                        device = "Pixel 2"
+                        apiLevel = 27
+                        systemImageSource = "aosp"
+                    }
+                }
+                groups {
+                    maybeCreate("api27and30").apply {
+                        targetDevices.add(devices["pixel2api30"])
+                        targetDevices.add(devices["pixel2api30atd"])
+                        targetDevices.add(devices["pixel2api27"])
+                    }
+                }
+            }
         }
     }
 
@@ -48,4 +82,17 @@ dependencies {
 
     testImplementation(libs.junit.jupiter)
     testImplementation(libs.assertj.core)
+    androidTestImplementation(libs.bundles.android.test)
+    androidTestImplementation(libs.assertj.core)
+
+    // Room
+    implementation(libs.room.runtime)
+    implementation(libs.room.ktx)
+    kapt(libs.room.compiler)
+
+    // Hilt
+    implementation(libs.hilt.android)
+    kapt(libs.hilt.compiler)
+    kaptAndroidTest(libs.hilt.compiler)
+    androidTestImplementation(libs.hilt.android.testing)
 }
