--- server/di/src/main/java/com/example/di/meal/MealApiModule.kt
@@ -1,18 +1,0 @@
-package com.example.di.meal
-
-import com.example.server.meal.api.MealApi
-import com.example.server.meal.api.mealApi
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object MealApiModule {
-
-    @Singleton
-    @Provides
-    fun provideMealApi(): MealApi = mealApi
-}
